update to Change-Id: I25a4d49f19e8067d114d56c100963d48d9d21e66
A new flag has been introduced "Config.enableReset". By default, it's set to true but if you would like the devices setup to remain after a page refresh you have to put this flag to false.
It works server-side and the existing devices are not reset but on the client side the refresh function is not triggered, function that should show up the existing devices.
I added the trigger for this function call as well as some necessary adjustments.